Rename Deserializers to Adapters
