+++ packages/create-django-render/template/client/src/adapters/ServerRenderedField.tsx
