+++ packages/create-django-render/template/client/src/adapters/widgets/base.tsx
